# Add Cache-Control headers on AP endpoints