This enables future use of Varnish in front of WriteFreely (see [T693](https://writefreely.org/tasks/693) for details).

Includes:

* ActivityPub Collection fetching via canonical URL
* ActivityPub Collection fetching via API
* ActivityPub Post fetching via canonical URL
* ActivityPub Post fetching via API

(By "ActivityPub fetching" I mean requests with the header `Accept: application/activity+json`)

This closes [T693](https://writefreely.org/tasks/693).