@robjloranger 2019-09-10
I'm not sure if you did any benchmarks but I did :grin: on the collection endpoint `domain.com/collection`

Without Varnish:
```
Transactions:		        1020 hits
Availability:		      100.00 %
Elapsed time:		        2.14 secs
Data transferred:	        0.71 MB
Response time:		        0.47 secs
Transaction rate:	      476.64 trans/sec
Throughput:		        0.33 MB/sec
Concurrency:		      222.43
Successful transactions:        1020
Failed transactions:	           0
Longest transaction:	        0.98
Shortest transaction:	        0.16
```

With Varnish:
```
Transactions:		        1533 hits
Availability:		      100.00 %
Elapsed time:		        3.36 secs
Data transferred:	        0.75 MB
Response time:		        0.41 secs
Transaction rate:	      456.25 trans/sec
Throughput:		        0.22 MB/sec
Concurrency:		      188.11
Successful transactions:        1533
Failed transactions:	           0
Longest transaction:	        0.71
Shortest transaction:	        0.15
```

That's just with two iterations of 255 concurrent users.