@thebaer 2019-09-11
Nice! Thanks for the thorough testing / benchmarking -- I hadn't done that :smile:. And to be sure, this only introduces caching on ActivityPub endpoints where we're doing `GET` requests?