@robjloranger 2019-09-11
Yep, the other requests include the `Via: Varnish` header but no age or cache control. Try `curl -IL rob.writefreely.dev` to see a non cached page