@robjloranger 2019-09-11
Got it, I just added a case in the subroutine for receive `sub vcl_recv`. When the request does not contain the accept header for activity json it passes the request direct to the backend. With the header you process through caching. To see it in action, including the cached version aging with non cache requests between, alternate the following a few times:

```
curl -IL rob.writefreely.dev/log/hello-hubzilla -H "Accept: application/activity+json"

curl -IL rob.writefreely.dev/log/hello-hubzilla
```