@thebaer 2020-01-31
Thanks, let's continue that discussion on writefreely/documentation#7 -- I think there are some other things we might want to tweak. For now, I'll merge this in.